+++ CommonAssemblyInfo.cs
@@ -2,11 +2,11 @@
 using System.Runtime.InteropServices;
 
 [assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("Comet Overdrive, Brett Story")]
+[assembly: AssemblyCompany("Macabresoft, Brett Story")]
 [assembly: AssemblyProduct("Macabre2D.Framework")]
 [assembly: AssemblyCopyright("Copyright © 2016 - 2020")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 [assembly: ComVisible(false)]
-[assembly: AssemblyVersion("0.3.14.0")]
-[assembly: AssemblyFileVersion("0.3.14.0")]
+[assembly: AssemblyVersion("0.3.15.0")]
+[assembly: AssemblyFileVersion("0.3.15.0")]